# server-src/models/requests/INewEntryParams.ts
export interface INewEntryParams {
    pounds: number;
    crop?: string;
    priceTotal?: number;
    harvester?: string;
    comments?: string;
    recipient?: string;
    selectedVariety?: string;
}