# server-src/swagger.json
{
	"basePath": "/api",
	"consumes": [
		"application/json"
	],
	"definitions": {
...
		"ICropVm": {
			"properties": {
				"name": {
					"type": "string"
				},
				"variety": {
					"type": "array",
					"items": {
						"type": "string"
					}
				},
				"pricePerPound": {
					"type": "number",
					"format": "double"
				},
				"createdOn": {
					"type": "string",
					"format": "date-time"
				},
				"updatedOn": {
					"type": "string",
					"format": "date-time"
				},
				"_id": {
					"type": "string"
				}
			},
			"required": [
				"name",
				"variety",
				"pricePerPound"
			],
			"type": "object"
		},
		"IHarvesterVm": {
			"properties": {
				"firstName": {
					"type": "string"
				},
				"lastName": {
					"type": "string"
				},
				"createdOn": {
					"type": "string",
					"format": "date-time"
				},
				"updatedOn": {
					"type": "string",
					"format": "date-time"
				},
				"_id": {
					"type": "string"
				}
			},
			"required": [
				"firstName",
				"lastName"
			],
			"type": "object"
		},
		"OrganizationType": {
			"enum": [
				"Purchased",
				"Donated",
				"Internal"
			],
			"type": "string"
		},
		"IOrganizationVm": {
			"properties": {
				"orgType": {
					"$ref": "#/definitions/OrganizationType"
				},
				"name": {
					"type": "string"
				},
				"createdOn": {
					"type": "string",
					"format": "date-time"
				},
				"updatedOn": {
					"type": "string",
					"format": "date-time"
				},
				"_id": {
					"type": "string"
				}
			},
			"type": "object"
		},
		"IEntryVm": {
			"properties": {
				"crop": {
					"$ref": "#/definitions/ICropVm"
				},
				"pounds": {
					"type": "number",
					"format": "double"
				},
				"priceTotal": {
					"type": "number",
					"format": "double"
				},
				"harvester": {
					"$ref": "#/definitions/IHarvesterVm"
				},
				"comments": {
					"type": "string"
				},
				"recipient": {
					"$ref": "#/definitions/IOrganizationVm"
				},
				"selectedVariety": {
					"type": "string"
				},
				"createdOn": {
					"type": "string",
					"format": "date-time"
				},
				"updatedOn": {
					"type": "string",
					"format": "date-time"
				},
				"_id": {
					"type": "string"
				}
			},
			"required": [
				"crop",
				"pounds",
				"priceTotal",
				"harvester",
				"comments",
				"recipient",
				"selectedVariety"
			],
			"type": "object"
		},
		"INewEntryParams": {
			"properties": {
				"pounds": {
					"type": "number",
					"format": "double"
				},
				"crop": {
					"type": "string"
				},
				"priceTotal": {
					"type": "number",
					"format": "double"
				},
				"harvester": {
					"type": "string"
				},
				"comments": {
					"type": "string"
				},
				"recipient": {
					"type": "string"
				},
				"selectedVariety": {
					"type": "string"
				}
			},
			"required": [
				"pounds"
			],
			"type": "object"
		},
...
		"INewCropParams": {
			"properties": {
				"name": {
					"type": "string"
				},
				"variety": {
					"type": "array",
					"items": {
						"type": "string"
					}
				},
				"pricePerPound": {
					"type": "number",
					"format": "double"
				}
			},
			"required": [
				"name",
				"variety",
				"pricePerPound"
			],
			"type": "object"
		},
...
	},
	"info": {
		"description": "CodeWithCause-NodeExpress Documentation",
		"version": "1.0.0",
		"title": "CodeWithCause-Node Express"
	},
	"paths": {
...
		"/entries/create": {
			"post": {
				"operationId": "RegisterEntry",
				"produces": [
					"application/json"
				],
				"responses": {
					"200": {
						"description": "Ok",
						"schema": {
							"$ref": "#/definitions/IEntryVm"
						}
					}
				},
				"tags": [
					"Entry"
				],
				"security": [],
				"parameters": [
					{
						"in": "body",
						"name": "newEntryParams",
						"required": true,
						"schema": {
							"$ref": "#/definitions/INewEntryParams"
						}
					}
				]
			}
		},
...
		"/entries/{id}": {
			"get": {
				"operationId": "GetSingleEntry",
				"produces": [
					"application/json"
				],
				"responses": {
					"200": {
						"description": "Ok",
						"schema": {
							"$ref": "#/definitions/IEntryVm"
						}
					}
				},
				"tags": [
					"Entry"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"type": "string"
					}
				]
			},
			"put": {
				"operationId": "UpdateEntry",
				"produces": [
					"application/json"
				],
				"responses": {
					"200": {
						"description": "Ok",
						"schema": {
							"$ref": "#/definitions/IEntryVm"
						}
					}
				},
				"tags": [
					"Entry"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"type": "string"
					},
					{
						"in": "body",
						"name": "updatedEntryParams",
						"required": true,
						"schema": {
							"$ref": "#/definitions/INewEntryParams"
						}
					}
				]
			},
			"delete": {
				"operationId": "DeleteEntry",
				"produces": [
					"application/json"
				],
				"responses": {
					"200": {
						"description": "Ok",
						"schema": {
							"$ref": "#/definitions/IEntryVm"
						}
					}
				},
				"tags": [
					"Entry"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"type": "string"
					}
				]
			}
		},
...
		"/crops/create": {
			"post": {
				"operationId": "RegisterCrop",
				"produces": [
					"application/json"
				],
				"responses": {
					"200": {
						"description": "Ok",
						"schema": {
							"$ref": "#/definitions/ICropVm"
						}
					}
				},
				"tags": [
					"Crop"
				],
				"security": [],
				"parameters": [
					{
						"in": "body",
						"name": "newCropParams",
						"required": true,
						"schema": {
							"$ref": "#/definitions/INewCropParams"
						}
					}
				]
			}
		},
...
		"/crops/{id}": {
			"get": {
				"operationId": "GetSingleCrop",
				"produces": [
					"application/json"
				],
				"responses": {
					"200": {
						"description": "Ok",
						"schema": {
							"$ref": "#/definitions/ICropVm"
						}
					}
				},
				"tags": [
					"Crop"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"type": "string"
					}
				]
			},
			"put": {
				"operationId": "UpdateCrop",
				"produces": [
					"application/json"
				],
				"responses": {
					"200": {
						"description": "Ok",
						"schema": {
							"$ref": "#/definitions/ICropVm"
						}
					}
				},
				"tags": [
					"Crop"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"type": "string"
					},
					{
						"in": "body",
						"name": "updateCropParams",
						"required": true,
						"schema": {
							"$ref": "#/definitions/INewCropParams"
						}
					}
				]
			},
			"delete": {
				"operationId": "DeleteCrop",
				"produces": [
					"application/json"
				],
				"responses": {
					"200": {
						"description": "Ok",
						"schema": {
							"$ref": "#/definitions/ICropVm"
						}
					}
				},
				"tags": [
					"Crop"
				],
				"security": [
					{
						"JWT": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"type": "string"
					}
				]
			}
		},
...
	},
	"produces": [
		"application/json"
	],
	"swagger": "2.0",
	"securityDefinitions": {
		"JWT": {
			"type": "apiKey",
			"name": "Authorization",
			"in": "header"
		}
	},
...
}